feat(run-contract): move the validator to lib/scripts/run_contract_validate.py and harden it for arbitrary JSON (#82A)
Non-object roots, lists where strings/enums are expected, numeric stop_detail,
unhashable worker names → field-naming errors, never TypeError/AttributeError.
tests/test_run_contract.py imports the production module. Structural
validation only; the rerun predicate is separate and the store never calls it.
